final  debug
 some littele change in view and controller

+++ first-project-zahra-mortazavi/src/model/goods/vehicles/Car.java
@@ -3,7 +3,7 @@ package model.goods.vehicles;
 import model.goods.GoodsCategoryModel;
 
 public class Car extends Vehicles {
-   final private int carEngineVolume;
+   final private double carEngineVolume;
     final   private boolean AutomaticCar;
 
     public double getCarEngineVolume() {
@@ -18,7 +18,7 @@ public class Car extends Vehicles {
 
 
 
-    public Car(String goodsName, double goodsPrice, int goodsInventory, String manufacturingCompany, int carEngineVolume, boolean automaticCar) {
+    public Car(String goodsName, double goodsPrice, int goodsInventory, String manufacturingCompany, double carEngineVolume, boolean automaticCar) {
         super(goodsName, goodsPrice, goodsInventory, manufacturingCompany);
         this.carEngineVolume = carEngineVolume;
         AutomaticCar = automaticCar;
